pyproject.toml
[project]
name = "thread-manager"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "atproto>=0.0.58",
    "azure-core>=1.32.0",
    "azure-cosmos>=4.9.0",
    "azure-monitor-opentelemetry>=1.6.4",
    "azure-storage-blob>=12.24.1",
    "beautifulsoup4>=4.13.3",
    "flask>=3.1.0",
    "flask-session>=0.8.0",
    "fredapi>=0.5.2",
    "hubspot-api-client>=11.1.0",
    "identity>=0.9.2",
    "matplotlib>=3.10.0",
    "openai>=1.61.0",
    "pillow>=11.1.0",
    "pybirdbuddy>=0.0.19",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "sendgrid>=6.11.0",
    "tenacity>=9.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]

[tool.pytest.ini_options]
pythonpath = ["."]
log_cli = true
log_cli_level = "DEBUG"
filterwarnings = [
    "ignore:Use 'content=<...>' to upload raw bytes/text content.:DeprecationWarning"
]
addopts = "-k 'None'"